SPLIT WORKSPACES || READ BELOW FOR MORE INFO|| HOPEFULLY THIS IS THE RIGHT FILE
There is a new SplitScreen class - Ignore that, It was a failure.
The way split workspaces works is that you type either /splitvert or /splithoriz in a canvas.  Then click the Split button under tools.  TADA.  Keep doing it!  Try doing both in the same pane!  NOW SWITCH THE ORDER OF THE COMMANDS.  gorgeous.  Outside the event code for the button there is some other code thrown about, all of it should be marked with //!!!NEW NEW NEW, but who knows, maybe I missed something.  Do the extra workspaces save? No.  Can you get rid of them?  Not unless you quit the app.  But can we change that? Y-E-S.  hit me up with questions when I'm awake.